pyproject.toml
[project]
name = "llmcord"
version = "0.1.0"
description = "llmcord+ transforms Discord into a collaborative LLM frontend. It works with practically any LLM, remote or locally hosted."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "httpx>=0.28.1",
    "openai>=1.99.5",
    "pyyaml>=6.0.2",
]

[tool.uv]
package = true

[project.scripts]
llmcord = "llmcord.bot:_run"

[dependency-groups]
dev = [
    "basedpyright==1.29.5", # currently downgraded due to https://github.com/DetachHead/basedpyright/issues/1395
    "ruff>=0.12.8",
]

[tool.basedpyright]
typeCheckingMode = "basic"
